+++ r4s/src/perception/package.xml
@@ -4,7 +4,7 @@
   <name>perception</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
-  <maintainer email="">mohsin</maintainer>
+  <maintainer email="mohsinalimirxa@gmail.com"></maintainer>
   <license>TODO: License declaration</license>
 
   <depend>sensor_msgs</depend>
