+++ r4s/src/perception/setup.py
@@ -1,27 +1,30 @@
-from setuptools import find_packages, setup
-import os
+from setuptools import setup
 from glob import glob
+import os
 
 package_name = 'perception'
 
 setup(
     name=package_name,
     version='0.0.0',
-    packages=find_packages(exclude=['test']),
+    packages=[package_name],
     data_files=[
         ('share/ament_index/resource_index/packages',
             ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
-        (os.path.join('share', package_name, 'launch'), glob(os.path.join('launch', '*launch.[pxy][yma]*'))),
-        (os.path.join('share', package_name, 'models'), glob('models/*.pt')),
+
+        # Install models
+        (os.path.join('share', package_name, 'models'), glob('models/*')),
+
+        # Install launch files
+        (os.path.join('share', package_name, 'launch'), glob('launch/*.py')),
     ],
     install_requires=['setuptools'],
     zip_safe=True,
     maintainer='mohsin',
     maintainer_email='mohsinalimirxa@gmail.com',
     description='TODO: Package description',
-    license='Apache-2.0',
-    tests_require=['pytest'],
+    license='TODO: License declaration',
     entry_points={
         'console_scripts': [
             "yolo_node = perception.yolo_object_detection:main",
