+++ requirements.txt
@@ -4,6 +4,8 @@ action-tutorials-interfaces==0.33.5
 action-tutorials-py==0.33.5
 actionlib-msgs==5.3.6
 actuator-msgs==0.0.1
+ament-clang-format==0.17.2
+ament-clang-tidy==0.17.2
 ament-cmake-test==2.5.4
 ament-copyright==0.17.2
 ament-cppcheck==0.17.2
@@ -17,26 +19,23 @@ ament-pep257==0.17.2
 ament-uncrustify==0.17.2
 ament-xmllint==0.17.2
 angles==1.16.0
-asttokens==3.0.0
 bond==4.1.2
 builtin-interfaces==2.0.2
-certifi==2025.4.26
-charset-normalizer==3.4.2
-comm==0.2.2
+camera-calibration==5.0.11
+certifi==2025.11.12
+charset-normalizer==3.4.4
 composition-interfaces==2.0.2
-contourpy==1.3.2
-control-msgs==5.4.0
-controller-manager==4.29.0
-controller-manager-msgs==4.29.0
+contourpy==1.3.3
+control-msgs==5.4.1
+controller-manager==4.32.0
+controller-manager-msgs==4.32.0
 cv-bridge==4.1.0
 cycler==0.12.1
-debugpy==1.8.14
-decorator==5.2.1
 demo-nodes-py==0.33.5
 diagnostic-msgs==5.3.6
-diagnostic-updater==4.2.4
+diagnostic-updater==4.2.6
 domain-coordinator==0.12.0
-dwb-msgs==1.3.6
+dwb-msgs==1.3.7
 example-interfaces==0.12.0
 examples-rclpy-executors==0.19.5
 examples-rclpy-minimal-action-client==0.19.5
@@ -45,90 +44,83 @@ examples-rclpy-minimal-client==0.19.5
 examples-rclpy-minimal-publisher==0.19.5
 examples-rclpy-minimal-service==0.19.5
 examples-rclpy-minimal-subscriber==0.19.5
-executing==2.2.0
-filelock==3.18.0
-fonttools==4.58.0
-fsspec==2025.3.2
-generate-parameter-library-py==0.4.0
+filelock==3.20.0
+fonttools==4.61.0
+fsspec==2025.12.0
+gazebo-msgs==3.8.0
+generate-parameter-library-py==0.5.0
 geographic-msgs==1.0.6
 geometry-msgs==5.3.6
-gps-msgs==2.0.4
-idna==3.10
+gps-msgs==2.1.0
+idna==3.11
 image-geometry==4.1.0
 interactive-markers==2.5.4
-ipykernel==6.29.5
-ipython==9.2.0
-ipython_pygments_lexers==1.1.1
+irobot-create-common-bringup==3.0.4
 irobot-create-msgs==3.0.0
-jedi==0.19.2
 Jinja2==3.1.6
 joint-state-publisher==2.4.0
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-kiwisolver==1.4.8
-laser-geometry==2.7.0
-launch==3.4.4
-launch-ros==0.26.7
-launch-testing==3.4.4
-launch-testing-ros==0.26.7
-launch-xml==3.4.4
-launch-yaml==3.4.4
+joint-state-publisher-gui==2.4.0
+kiwisolver==1.4.9
+laser-geometry==2.7.1
+launch==3.4.5
+launch-param-builder==0.1.1
+launch-ros==0.26.8
+launch-testing==3.4.5
+launch-testing-ros==0.26.8
+launch-xml==3.4.5
+launch-yaml==3.4.5
 lifecycle-msgs==2.0.2
 logging-demo==0.33.5
 map-msgs==2.4.1
-MarkupSafe==3.0.2
-matplotlib==3.10.3
-matplotlib-inline==0.1.7
+MarkupSafe==3.0.3
+matplotlib==3.10.8
 message-filters==4.11.6
+moveit-configs-utils==2.12.3
+moveit-msgs==2.6.0
 mpmath==1.3.0
-my-robot-controller==0.0.0
-nav-2d-msgs==1.3.6
+nav-2d-msgs==1.3.7
 nav-msgs==5.3.6
-nav2-common==1.3.6
-nav2-msgs==1.3.6
+nav2-common==1.3.7
+nav2-loopback-sim==1.0.0
+nav2-msgs==1.3.7
 nav2-simple-commander==1.0.0
-nest-asyncio==1.6.0
-networkx==3.4.2
+networkx==3.6.1
 numpy==1.26.4
-nvidia-cublas-cu12==12.6.4.1
-nvidia-cuda-cupti-cu12==12.6.80
-nvidia-cuda-nvrtc-cu12==12.6.77
-nvidia-cuda-runtime-cu12==12.6.77
-nvidia-cudnn-cu12==9.5.1.17
-nvidia-cufft-cu12==11.3.0.4
-nvidia-cufile-cu12==1.11.1.6
-nvidia-curand-cu12==10.3.7.77
-nvidia-cusolver-cu12==11.7.1.2
-nvidia-cusparse-cu12==12.5.4.2
-nvidia-cusparselt-cu12==0.6.3
-nvidia-nccl-cu12==2.26.2
-nvidia-nvjitlink-cu12==12.6.85
-nvidia-nvtx-cu12==12.6.77
-opencv-python==4.11.0.86
+nvidia-cublas-cu12==12.8.4.1
+nvidia-cuda-cupti-cu12==12.8.90
+nvidia-cuda-nvrtc-cu12==12.8.93
+nvidia-cuda-runtime-cu12==12.8.90
+nvidia-cudnn-cu12==9.10.2.21
+nvidia-cufft-cu12==11.3.3.83
+nvidia-cufile-cu12==1.13.1.3
+nvidia-curand-cu12==10.3.9.90
+nvidia-cusolver-cu12==11.7.3.90
+nvidia-cusparse-cu12==12.5.8.93
+nvidia-cusparselt-cu12==0.7.1
+nvidia-nccl-cu12==2.27.5
+nvidia-nvjitlink-cu12==12.8.93
+nvidia-nvshmem-cu12==3.3.20
+nvidia-nvtx-cu12==12.8.90
+object-recognition-msgs==2.0.0
+octomap-msgs==2.0.1
+opencv-python==4.12.0.88
 osrf-pycommon==2.1.6
 packaging==25.0
 pal-statistics==2.6.2
 pal-statistics-msgs==2.6.2
-pandas==2.2.3
-parso==0.8.4
 pcl-msgs==1.0.0
 pendulum-msgs==0.33.5
-pexpect==4.9.0
-pillow==11.2.1
-platformdirs==4.3.8
-prompt_toolkit==3.0.51
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-py-cpuinfo==9.0.0
-Pygments==2.19.1
-pyparsing==3.2.3
+pillow==12.0.0
+polars==1.36.1
+polars-runtime-32==1.36.1
+psutil==7.1.3
+py-trees==2.3.0
+py-trees-ros==2.3.0
+py-trees-ros-interfaces==2.1.1
+pyparsing==3.2.5
 python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
 python-qt-binding==2.2.1
-pytz==2025.2
-PyYAML==6.0.2
-pyzmq==26.4.0
+PyYAML==6.0.3
 qt-dotgraph==2.7.5
 qt-gui==2.7.5
 qt-gui-cpp==2.7.5
@@ -137,23 +129,25 @@ quality-of-service-demo-py==0.33.5
 rcl-interfaces==2.0.2
 rclpy==7.1.4
 rcutils==6.7.2
-realsense2-camera-msgs==4.55.1
-requests==2.32.3
+realsense2-camera-msgs==4.56.4
+requests==2.32.5
 resource-retriever==3.4.4
 rmw-dds-common==3.1.0
 robot-localization==3.8.2
-ros-gz-bridge==1.0.12
-ros-gz-interfaces==1.0.12
-ros-gz-sim==1.0.12
+ros-gz-bridge==1.0.15
+ros-gz-interfaces==1.0.15
+ros-gz-sim==1.0.15
+ros2-controllers-test-nodes==4.27.0
 ros2action==0.32.4
 ros2bag==0.26.7
 ros2bag-mcap-cli==0.26.7
 ros2bag-sqlite3-cli==0.26.7
 ros2cli==0.32.4
 ros2component==0.32.4
+ros2controlcli==4.32.0
 ros2doctor==0.32.4
 ros2interface==0.32.4
-ros2launch==0.26.7
+ros2launch==0.26.8
 ros2lifecycle==0.32.4
 ros2multicast==0.32.4
 ros2node==0.32.4
@@ -200,49 +194,61 @@ rqt-service-caller==1.2.1
 rqt-shell==1.2.2
 rqt-srv==1.2.2
 rqt-topic==1.7.3
-scipy==1.15.3
+scenario-execution==1.3.0
+scenario-execution-control==1.3.0
+scenario-execution-coverage==1.3.0
+scenario-execution-gazebo==1.3.0
+scenario-execution-interfaces==1.3.0
+scenario-execution-nav2==1.3.0
+scenario-execution-os==1.3.0
+scenario-execution-py-trees-ros==1.2.0
+scenario-execution-ros==1.3.0
+scenario-execution-x11==1.3.0
+scipy==1.16.3
 sensor-msgs==5.3.6
 sensor-msgs-py==5.3.6
 service-msgs==2.0.2
-setuptools==80.7.1
+setuptools==80.9.0
 shape-msgs==5.3.6
+simple-launch==1.11.0
 six==1.17.0
 slam-toolbox==2.8.3
 smclib==4.1.2
+srdfdom==2.0.7
 sros2==0.13.3
-stack-data==0.6.3
 statistics-msgs==2.0.2
 std-msgs==5.3.6
 std-srvs==5.3.6
 stereo-msgs==5.3.6
 sympy==1.14.0
 teleop-twist-keyboard==2.4.0
-tf2-geometry-msgs==0.36.11
-tf2-kdl==0.36.11
-tf2-msgs==0.36.11
-tf2-py==0.36.11
-tf2-ros-py==0.36.11
-tf2-sensor-msgs==0.36.11
-tf2-tools==0.36.11
+tf-transformations==1.1.0
+tf2-geometry-msgs==0.36.12
+tf2-kdl==0.36.12
+tf2-msgs==0.36.12
+tf2-py==0.36.12
+tf2-ros-py==0.36.12
+tf2-sensor-msgs==0.36.12
+tf2-tools==0.36.12
 theora-image-transport==4.0.4
 topic-monitor==0.33.5
-torch==2.7.0
-torchvision==0.22.0
-tornado==6.5
-tqdm==4.67.1
-traitlets==5.14.3
+topic-tools==1.3.3
+topic-tools-interfaces==1.3.3
+torch==2.9.1
+torchvision==0.24.1
 trajectory-msgs==5.3.6
 transforms3d==0.4.2
-triton==3.3.0
+triton==3.5.1
+turtlebot4-msgs==2.1.0
+turtlebot4-navigation==2.1.0
 turtlesim==1.8.3
 type-description-interfaces==2.0.2
-typing_extensions==4.13.2
-tzdata==2025.2
-ultralytics==8.3.140
-ultralytics-thop==2.0.14
+typing_extensions==4.15.0
+ultralytics==8.3.236
+ultralytics-thop==2.0.18
 unique-identifier-msgs==2.5.0
-urllib3==2.4.0
+urdfdom-py==1.2.1
+urllib3==2.6.2
 vision-msgs==4.1.1
 visualization-msgs==5.3.6
-wcwidth==0.2.13
 xacro==2.0.13
